+++ src/clients/__init__.py
@@ -1,2 +1,2 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
-# Copyright (C) 2024 Wewoc (github.com/wewoc)
+# Copyright (C) 2026 Wewoc (github.com/wewoc)