+++ src/tests/conftest.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
-# Copyright (C) 2024 Wewoc (github.com/wewoc)
+# Copyright (C) 2026 Wewoc (github.com/wewoc)
 
 """
 tests/conftest.py